Update to dnsjava 1.6.2, which is BSD licensed
git-svn-id: https://svn.apache.org/repos/asf/james/server@109087 13f79535-47bb-0310-9956-ffa450edef68

+++ branches/branch_2_1_fcs/src/java/org/apache/james/dnsserver/DNSServer.java
@@ -67,7 +67,7 @@ public class DNSServer
     /**
      * Whether the DNS response is required to be authoritative
      */
-    private byte dnsCredibility;
+    private int dnsCredibility;
 
     /**
      * The DNS servers to be used by this service
